fix printf size_t problem with "%lu" and (unsigned long) cast.
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@2923 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/common/main.c
@@ -119,9 +119,7 @@ handle(char *progname,
     {
       fprintf(stderr,
               "%s: %lu bytes copied from `%s' to `%s'\n",
-              progname,
-              bytes_copied,
-              infilename,
+              (unsigned long) progname, bytes_copied, infilename,
               outfilename);
     }
   return 0;
